LayoutTests/editing/selection/update-selection-with-find-dissociates-range-expected.txt

This tests calling window.find twice. getSelection().getRangeAt(0) should return distinct ranges.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS range1 != range2 is true
PASS successfullyParsed is true

TEST COMPLETE
hello, world
